Move date conversion logic to route rather than model
