+++ backend/src/models/event.test.ts
@@ -25,7 +25,7 @@ describe("create", function () {
     buskerId: 0,
     title: "Test title",
     type: "Test type",
-    date: "2023-10-12",
+    date: "Tue, Oct 10, 2023",
     startTime: "01:16",
     endTime: "02:16",
     coordinates: { lat: 1, lng: 1 },
@@ -37,7 +37,6 @@ describe("create", function () {
     let event = await Event.create(newEvent);
     expect(event).toEqual({
       ...newEvent,
-      date: "Thu, Oct 12, 2023",
       buskerId: testBuskerIds[0],
       id: testEventIds[0] + 1,
     });
@@ -112,7 +111,7 @@ describe("getAll", function () {
     buskerId: 0,
     title: "Test title",
     type: "Test type",
-    date: "2023-10-12",
+    date: "Tue, Oct 10, 2023",
     startTime: "02:16",
     endTime: "03:16",
     coordinates: { lat: 1, lng: 1 },
@@ -141,7 +140,7 @@ describe("get", function () {
       buskerName: testBuskerNames[0],
       title: "e1",
       type: "E1",
-      date: "2023-10-10",
+      date: "Tue, Oct 10, 2023",
       endTime: "14:00",
       startTime: "13:00",
       coordinates: { lat: 0, lng: 0 },
@@ -165,7 +164,7 @@ describe("update", function () {
     buskerId: 0,
     title: "newTitle",
     type: "NewType",
-    date: "2023-10-11",
+    date: "Wed, Oct 11, 2023",
     startTime: "00:00",
     endTime: "10:00",
     coordinates: { lat: 45, lng: 45 },
