+++ backend/src/routes/_testCommon.ts
@@ -67,7 +67,7 @@ export async function commonBeforeAll() {
       buskerId: testBuskerIds[0],
       title: "test event",
       type: "test type",
-      date: "2023-10-10",
+      date: "Tue, Oct 10, 2023",
       startTime: "13:00",
       endTime: "14:00",
       coordinates: { lat: 0, lng: 0 },
